fix: address second code review - simplify loops, fix fp comparison, clean up redundant expressions
Agent-Logs-Url: https://github.com/bherila/2025-website/sessions/c636b6a4-6bc3-4fa1-a057-6fa3079b01bd

+++ app/GenAiProcessor/Services/GenAiJobDispatcherService.php
@@ -1139,9 +1139,9 @@ PROMPT;
             if (is_bool($raw)) {
                 $boolValue = $raw;
             } elseif (is_int($raw) || is_float($raw)) {
-                if ((float) $raw === 1.0) {
+                if ((int) $raw === 1) {
                     $boolValue = true;
-                } elseif ((float) $raw === 0.0) {
+                } elseif ((int) $raw === 0) {
                     $boolValue = false;
                 }
             } elseif (is_string($raw)) {
